requirements.txt
decorator==4.4.2
Flask==1.1.2
fsspec==0.8.0
html5lib==1.1
ipython==5.8.0
ipython-genutils==0.2.0
jsonfield==3.1.0
mysql-connector==2.2.9
requests==2.22.0
six==1.15.0
xlrd==1.2.0
XlsxWriter==1.2.8
xmltodict==0.12.0
